requirements.txt
appdirs==1.4.3
appnexus==1.0
asn1crypto==0.22.0
azure-common==1.1.5
azure-nspkg==2.0.0
azure-storage==0.34.0
cffi==1.10.0
cryptography==1.8.1
enum34==1.1.6
futures==3.1.1
idna==2.5
ipaddress==1.0.18
packaging==16.8
pkg-resources==0.0.0
pycparser==2.17
pyparsing==2.2.0
python-dateutil==2.6.0
requests==2.13.0
six==1.10.0
boto3==1.4.7
